Finished Problem 4
Turns out the code that checked the number was not wrong, I just needed
to remove the cout from the Check function so that it wouldn't print out
every number that was checked. The program does take a couple minutes to
complete, but i think that is my programming skills limiting the speed
in which it happens. - Isaac