Finished Problem 5
I figured out that you don't need to check all the numbers, just the
ones that don't have multiples. I created an array to hold the numbers
to check. I then checked each number against all the numbers in the
array. If one of the numbers didn't match it was skipped instantly. I
also went up by 20 each time because the number has to be a multiple of
20. -Isaac