Finished Problem 6
Probably one of the easiest I've done so far. I have a function that
calculates the sum of the squares of all numbers up to 100, and then I
have a function to find the sum of all numbers up to 100 and then
squared it. I then have a function that subtracts them. This could have
been done easier but I decided to split up everything when it died. -
Isaac