Finished Problem 10
I finally finished Problem 10. I had to remember to set the for loop in
the function to start at 2 AND only go until its less than but not equal
to. I also changed the function to a bool instead of a void because that
was wrong. I probably won't be able to do Problem 11 but I might try
Problem 12 later. But before I do anything I'm going to add comments to
all my old code as that is important.
-Isaac