Finished first 25 problems
Finally finished the first 25 problems. All the code is mine, but some
of the equations used are ones where I had to go to the internet to
find, as I don't have the math skills for them. I learned a lot so far
and will probably now try to finish the first 50 problems.